ui/requirements.txt
# Streamlit UI Dependencies
streamlit>=1.28.0

# HTTP requests
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.2

# Environment variables
python-dotenv>=1.0.0

